nim backend : fix fs natives to behave like other backends

+++ tools/flowc/backends/nim/natives/fs/filesystem/deleteFile.nim
@@ -2,8 +2,11 @@
 import os
 
 proc deleteFile*(dir : string): string =
+  if (fileExists(dir)):
     try:
-        os.removeFile(dir)
-        return ""
+      os.removeFile(dir)
+      return ""
     except OSError as e:
-        return e.msg
+      return e.msg
+  else:
+    return "The system cannot find the path specified."
